+++ tests/apolloQueryObservable.ts
@@ -1,49 +1,120 @@
 import {
-  Observable,
-} from 'rxjs/Observable';
-
-import {
   ApolloQueryObservable,
 } from '../src/apolloQueryObservable';
 
+import {
+  mockClient,
+} from './_mocks';
+
+import gql from 'graphql-tag';
+
 import 'rxjs/add/operator/map';
 
-class ObservableQuery<T> extends Observable<T> {
-  refetch(v?: any) {}
-  stopPolling() {}
-  startPolling(n?: any) {}
-}
+const query = gql`
+  query heroes {
+    allHeroes {
+      heroes {
+        name
+      }
+    }
+  }
+`;
+const data = {
+  allHeroes: {
+    heroes: [{ name: 'Mr Foo' }, { name: 'Mr Bar' }],
+  },
+};
 
 describe('ApolloQueryObservable', () => {
-  it('should be able to call refetch()', () => {
-    const obs = new ObservableQuery();
-    const res = new ApolloQueryObservable(obs);
-    const variables = { foo: true };
+  let obsApollo;
+  let obsQuery;
+  let client;
 
-    spyOn(obs, 'refetch').and.returnValue('refetch');
-
-    expect(res.refetch(variables)).toEqual('refetch');
-    expect(obs.refetch).toHaveBeenCalledWith(variables);
+  beforeEach(() => {
+    client = mockClient({
+      request: { query },
+      result: { data },
+    });
+    obsApollo = client.watchQuery({ query });
+    obsQuery = new ApolloQueryObservable(obsApollo);
   });
 
-  it('should be able to call stopPolling()', () => {
-    const obs = new ObservableQuery();
-    const res = new ApolloQueryObservable(obs);
+  describe('regular', () => {
+    it('should be able to subscribe', () => {
+      expect(() => {
+        obsQuery.subscribe({
+          next() {},
+        });
+      }).not.toThrow();
+    });
 
-    spyOn(obs, 'stopPolling').and.returnValue('stopPolling');
+    it('should be able to receive data', (done) => {
+      obsQuery.subscribe({
+        next(result) {
+          expect(result.data).toEqual(data);
+          done();
+        }
+      });
+    });
 
-    expect(res.stopPolling()).toEqual('stopPolling');
-    expect(obs.stopPolling).toHaveBeenCalled();
+    it('should be able to receive an error', (done) => {
+      obsQuery.subscribe({
+        next() {},
+      });
+
+      obsQuery.subscribe({
+        error(error) {
+          expect(error instanceof Error).toBe(true);
+          done();
+        },
+      })
+    });
+
+    it('should be able to use a operator', (done) => {
+      obsQuery.map(result => result.data).subscribe({
+        next(result) {
+          expect(result).toEqual(data);
+          done();
+        },
+      });
+    });
   });
 
-  it('should be able to call custom method after operator', () => {
-    const obs = new ObservableQuery();
-    spyOn(obs, 'refetch').and.returnValue('refetch');
+  describe('apollo-specific', () => {
+    it('should be able to refech', () => {
+      spyOn(obsApollo, 'refetch').and.returnValue('promise');
 
-    const res = new ApolloQueryObservable(obs).map((i) => i);
+      const arg = 'foo';
+      const result = obsQuery.refetch(arg);
 
+      expect(obsApollo.refetch).toHaveBeenCalledWith(arg);
+      expect(result).toBe('promise');
+    });
 
-    expect(res['refetch']()).toEqual('refetch');
-    expect(obs.refetch).toHaveBeenCalled();
+    it('should be able to startPolling', () => {
+      spyOn(obsApollo, 'startPolling');
+
+      const arg = 200;
+      obsQuery.startPolling(arg);
+
+      expect(obsApollo.startPolling).toHaveBeenCalledWith(arg);
+    });
+
+    it('should be able to stopPolling', () => {
+      spyOn(obsApollo, 'stopPolling');
+
+      obsQuery.stopPolling();
+
+      expect(obsApollo.stopPolling).toHaveBeenCalled();
+    });
+
+    it('should be able to fetchMore', () => {
+      spyOn(obsApollo, 'fetchMore');
+
+      const arg = 200;
+      obsQuery.fetchMore(arg);
+
+      expect(obsApollo.fetchMore).toHaveBeenCalledWith(arg);
+    });
   });
 });